fix(commit-convention): exported the commit-convention functionality through the public api
