+++ test/integration/features/step_definitions/common-steps.js
@@ -16,5 +16,5 @@ When('the project is scaffolded', async function () {
     node_modules: stubbedNodeModules
   });
 
-  await scaffold({projectRoot: process.cwd()});
+  await scaffold({projectRoot: process.cwd(), configs: {}});
 });
